+++ transport_frames/indicators/indicator_terr.py
@@ -137,7 +137,7 @@ def indicator_territory(graph: nx.MultiDiGraph, territory: gpd.GeoDataFrame, ser
     def calculate_distances(from_gdf, to_gdf, weight='length_meter', unit_div=1000):
         if to_gdf is None:
             return None
-        return round(get_adj_matrix_gdf_to_gdf(from_gdf, to_gdf, graph, weight=weight, dtype=np.float64).min(axis=1) / unit_div, 3)
+        return round(get_adj_matrix_gdf_to_gdf(from_gdf, to_gdf, graph, weight=weight, dtype=np.float16).min(axis=1) / unit_div, 3)
 
     # Calculate distances to region admin center and region 1 centers
     if region_admin_center is None:
